software/server/project_mqtt2mysql/app/requirements.txt
paho-mqtt
mysql-connector-python
